hamming_code.py
import numpy as np


class HammingCode:
    """
    The HammingCode class implements encoding, error correction, and decoding
    of binary data using Hamming codes for various configurations.

    Attributes:
        m (int): Length of the data (number of information bits).
        n (int): Length of the code (total number of bits in the encoded word).
    """

    def __init__(self, m):
        """
        Initializes a HammingCode object based on the number of data bits.

        Args:
            m (int): Number of data bits, e.g., 4 for (7,4) code or 11 for (15,11).
        """
        self.m = m
        self.n = 2 ** np.ceil(np.log2(m + np.ceil(np.log2(m)) + 1)) - 1
        self.r = int(np.log2(self.n + 1))  # Number of parity bits

    def encode(self, data):
        """
        Encodes the given binary data using the Hamming code algorithm.

        Args:
            data (np.array): A numpy array containing the binary data (information bits).

        Returns:
            np.array: The encoded array of bits, including data and parity bits.
        """
        #TODO: Example encoding for Hamming code - insert appropriate logic
        pass

    def decode(self, encoded_data):
        """
        Decodes the encoded binary data and corrects any single-bit errors.

        Args:
            encoded_data (np.array): A numpy array of encoded bits (data + parity bits).

        Returns:
            np.array: The original binary data after decoding and correcting errors.
        """
        #TODO: Example decoding - insert appropriate error detection and decoding logic
        pass

    def correct_error(self, encoded_data):
        """
        Detects and corrects single-bit errors in the encoded binary data.

        Args:
            encoded_data (np.array): A numpy array of encoded bits (data + parity bits).

        Returns:
            np.array: The encoded data with corrected errors.
        """
        #TODO: Error detection and correction logic - insert appropriate code
        pass
